wp-a/requirements.txt
flask
flask-wtf
hashids
uwsgi
gitpython
